pyproject.toml
[tool.poetry]
name = "conpar"
version = "0.1.0"
description = "Configure and initialize Python Classes using JSON and other files."
authors = ["Johannes Hecher <hhecher@gmx.at>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.24.36"
azure-storage-blob = "^12.13.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.6.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
